cleaning up editor state
