+++ app/modules/crono/pure/editorToCrono.tsx
@@ -1,5 +1,5 @@
 import { CronoModeEnum, SetModeEnum } from '../../editor/enums';
-import { EditorStateType, TableSetType } from '../../editor/redux/editorTypes';
+import { EditorStateType, TableSetType } from '../../editor/editorTypes';
 import { CronoRunningType, CronoSetType, CronoStateType } from '../redux/CronoTypes';
 
 function initSet(originalSet: TableSetType): CronoSetType {
