+++ app/modules/editor/editorTypes.ts
@@ -1,5 +1,5 @@
-import { FixMe } from '../../../types';
-import { TableTypeEnum, SetTypeEnum } from '../enums';
+import { ImmutableJSType } from '../../redux/types';
+import { TableTypeEnum, SetTypeEnum } from './enums';
 
 export type EditorStateType = {
   trainingTable: TableType;
@@ -10,7 +10,7 @@ export type TableType = {
   base: number;
   type: TableTypeEnum;
   duration: number;
-  // endurance
+  // endurance attributes
   baseBreaks?: number;
   enduranceLaps?: number;
 };
@@ -21,17 +21,7 @@ export type TableSetType = {
   pos: number;
 };
 
-/**
- * TODO: Remove immutable js
- */
-export type ImmutableJSType = {
-  get: (prop: string) => FixMe;
-  getIn: (stack: string[]) => FixMe;
-  setIn: (stack: string[], data: FixMe) => FixMe;
-  update: (stack: string[] | string, cb: (data: FixMe) => FixMe) => FixMe;
-  toJS: () => EditorStateType;
-};
-
+// TODO: Removing immubtale
 export type ImmutableJSEditorStateType = ImmutableJSType;
 export type ImmutableJSEditorType = ImmutableJSType;
 export type ImmutableTrainingTableType = ImmutableJSType;
