+++ app/routes/Crono/CronoScene.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import SceneWrapper from '../../components/SceneWrapper/SceneWrapper';
 import { AppScreenType, RoutesEnum } from '../Routes';
 import CronoPane from './components/CronoPane/CronoPane';
-import { EditorStateType } from '../../modules/editor/redux/editorTypes';
+import { EditorStateType } from '../../modules/editor/editorTypes';
 
 export type CronoSceneParamList = { initialData: EditorStateType };
