+++ app/routes/EnduranceEditor/components/EnduranceForm/EnduranceForm.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect } from 'react';
 import CronoStartButton from '../../../../components/CronoStartButton/CronoStartButton';
-import { ImmutableJSEditorSetType } from '../../../../modules/editor/redux/editorTypes';
+import { ImmutableJSEditorSetType } from '../../../../modules/editor/editorTypes';
 import { Wrapper } from './EnduranceForm.styled';
 import EnduranceMainForm from './EnduranceMainForm';
 import { useAppDispatch, useAppSelector } from '../../../../redux/hooks';
