+++ app/routes/StaticEditor/components/StaticSetsList/StaticSet.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import EditorTimerInput from '../StaticFormInputs/EditorTimerInput';
-import { ImmutableJSEditorSetType } from '../../../../modules/editor/redux/editorTypes';
+import { ImmutableJSEditorSetType } from '../../../../modules/editor/editorTypes';
 
 type Props = {
   set: ImmutableJSEditorSetType;
